Re-enable suspending main loop when application is paused on android. Fixes #2328.
Previous behaviour allowed main loop to spin (at 10fps, as if window is minimized) while application was in background. This is both bad for battery life and also introduces a race condition where rapidly suspending and resuming application may get it into the state where application is resumed while some or most GPU objects are not recreated.